Move more functions to separate files
